requirements.txt
alembic==0.9.9
aniso8601==3.0.0
blinker==1.4
click==6.7
dominate==2.3.1
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cache==0.13.1
Flask-CKEditor==0.3.3
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Moment==0.6.0
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SQLAlchemy==2.3.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
mysqlclient==1.3.12
Pillow==5.1.0
python-dateutil==2.7.3
python-editor==1.0.3
pytz==2018.4
redis==2.10.6
six==1.11.0
SQLAlchemy==1.2.7
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.1
